feat(testing): move tests into subpackages
